# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "BuildingPulse API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/sponsor/cta": {
      "get": {
        "operationId": "cta_api_sponsor_cta_get",
        "parameters": [
          {
            "in": "query",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Prediction Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scenario_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cta"
      },
      "head": {
        "operationId": "cta_api_sponsor_cta_get",
        "parameters": [
          {
            "in": "query",
            "name": "prediction_id",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Prediction Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scenario_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cta"
      }
    }
  }
}